Manually committing empty dist/pl/js/components.js
As part of 93ef9603721f5444241220f2001f56349f6e84ee I removed the last
input to pl:js which causes it to skip regeneration of components.js.

I wasn't able to find a way to have it regenerate the file even with
empty inputs, but this approach should at least have it regenerate if
a new JS file is ever added.

I also considered removing the pl:js generation altogether, but there
are already dependencies on the output, components.js, and it seems
likely that we will want to readd JS in the future.